Update: Modifications in mStat_generate_report_pair Function
🔍 **Overview:**
This update involves significant modifications in the `mStat_generate_report_pair` function to enhance flexibility and usability. Users should note the change in parameters and their settings to adapt to the new configuration for generating reports.

🛠️ **Changes Made:**
- Introduced new variables `test.adj.vars`, `vis.adj.vars`, `feature.change.func`, `vis.feature.level`, `test.feature.level`, `bar.area.feature.no`, `heatmap.feature.no`, `dotplot.feature.no`, `feature.mt.method`, and `feature.sig.level`.
- Removed `adj.vars`, `alpha.obj`, `dist.obj`, and `change.func` variables.
- Updated default values for `alpha.name`, `base.size`, and other parameters for enhanced performance.
- `data(peerj32.obj)` is now called before invoking the `mStat_generate_report_pair` function.

⚙️ **Impact:**
- Provides users with more control and flexibility over report generation.
- Ensures more clear and concise report configuration and generation.

📘 **Documentation Update:**
An example of how to use the updated `mStat_generate_report_pair` function is provided in the documentation, ensuring users can easily adapt to the new settings and configuration.

We look forward to feedback and suggestions to continuously improve the user experience and functionality.

#Update #mStat #ReportGeneration