Enhancement of Documentation for `generate_taxa_test_long()` and Introduction of New Functionality
This push significantly **enhances the documentation** for `generate_taxa_test_long()`, providing a clearer, more detailed, and comprehensive guide for users wishing to analyze taxonomic changes in longitudinal microbiome studies. The update ensures that users are well-equipped to utilize this powerful tool for differential abundance testing between groups and timepoints.

**Key Highlights:**

- **Expanded Explanation:** Detailed insight into the longitudinal trend test on microbiome data, enabling users to effectively discern changes in microbial taxa abundance over time and/or in response to different experimental or observational groups.
- **Robust Examples:** The updated documentation includes robust examples, ensuring users can easily understand and implement the function for their specific needs.
- **Clear Outputs:** Explanation of the detailed results table output, which highlights significant taxonomic differences between groups over time.

In addition to the enhancement of existing documentation, this push introduces a **new functionality**, the `generate_taxa_volatility_test_long()` function. This additional tool calculates the volatility of taxa abundances in longitudinal microbiome data, offering users further insight into the dynamics of microbial populations.

- **Volatility Calculation:** Enables the calculation of the mean absolute difference in abundance between consecutive time points, normalized by the time difference.
- **Enhanced Understanding:** Assists in understanding the association between abundance volatility and grouping variables, providing a more accurate representation of microbiome data dynamics.

These comprehensive updates and additions bolster the utility and user-friendliness of MicrobiomeStat, ensuring it remains an invaluable resource for researchers exploring the complexities of microbial communities. The push reinforces our commitment to providing robust, clear, and accessible tools for the microbiome research community.