Acknowledgements Update: Expressing Gratitude to Dependencies and Pioneers
In this update, we've added a section in the MicrobiomeStat tutorial to express our gratitude towards the developers of the packages that MicrobiomeStat depends on. Additionally, we pay homage to the pioneering work done by the developers of other notable microbiome analysis packages. This recognition reflects the collaborative spirit of the microbiome research community and underscores the collective effort in advancing the field.